Split Tomcat into Base and Home
Previously the Tomcat contribution put everything into a single layer.  This
was unnecessary and led to rebuilds for changes that weren't likely to be
needed because dependencies changed even when Tomcat didn't.

This change splits the installation into CATALINA_HOME (the Tomcat binaries)
and CATALINA_BASE (the configuration and application symlink).  In addition,
this change adds support for optional externalized configuration.

Signed-off-by: Ben Hale <bhale@pivotal.io>