External Configuration Environment Variables
Previously, if external configuration needed to be included, the buildpack had
to be forked and an entry added to buildpack.toml.  Given that we feel that
forking isn't ideal, there should have been another way to contribute external
configuration in a less invasive way.

This change updates the buildpack to allow a user to set the
BP_TOMCAT_EXT_CONF_VERSION, BP_TOMCAT_EXT_CONF_URI, and
BP_TOMCAT_EXT_CONF_SHA256 environment variables to point at an external
configuration during the detect and build phases.

Signed-off-by: Ben Hale <bhale@pivotal.io>